Fixing linter errors
