+++ src/assets/styles/Kenya.css
@@ -70,60 +70,6 @@
   color: #1b1b1c;
 }
 
-#phone-input {
-  background-color: yellow;
-}
-.react-international-phone-input {
-  width: 100;
-  background-color: yellow;
-  height: 45px;
-}
-
-/* Price Button */
-.price-button {
-  background-color: #004a7f;
-  -webkit-border-radius: 10px;
-  border-radius: 10px;
-  border: none;
-  color: #fff;
-  cursor: pointer;
-  display: inline-block;
-  font-size: 20px;
-  padding: 5px 10px;
-  text-align: center;
-  text-decoration: none;
-  -webkit-animation: glowing 1500ms infinite;
-  -moz-animation: glowing 1500ms infinite;
-  -o-animation: glowing 1500ms infinite;
-  animation: glowing 1500ms infinite;
-}
-
-/* stylelint-disable declaration-block-single-line-max-declarations */
-@-webkit-keyframes glowing {
-  0% { background-color: #669b45; -webkit-box-shadow: 0 0 3px #669b45; }
-  50% { background-color: #57d309; -webkit-box-shadow: 0 0 40px #57d309; }
-  100% { background-color: #669b45; -webkit-box-shadow: 0 0 3px #669b45; }
-}
-
-@-moz-keyframes glowing {
-  0% { background-color: #669b45; -moz-box-shadow: 0 0 3px #669b45; }
-  50% { background-color: #57d309; -moz-box-shadow: 0 0 40px #57d309; }
-  100% { background-color: #669b45; -moz-box-shadow: 0 0 3px #669b45; }
-}
-
-@-o-keyframes glowing {
-  0% { background-color: #669b45; box-shadow: 0 0 3px #669b45; }
-  50% { background-color: #57d309; box-shadow: 0 0 40px #57d309; }
-  100% { background-color: #669b45; box-shadow: 0 0 3px #669b45; }
-}
-
-@keyframes glowing {
-  0% { background-color: #669b45; box-shadow: 0 0 3px #669b45; }
-  50% { background-color: #57d309; box-shadow: 0 0 40px #57d309; }
-  100% { background-color: #669b45; box-shadow: 0 0 3px #669b45; }
-}
-/* stylelint-enable declaration-block-single-line-max-declarations */
-
 /**
 ##############################################
 ############ COAST TOURS  ####################
